+++ src/pages/screening/student/StudentHistory.tsx
@@ -1,18 +1,19 @@
 import { Typography } from '@/components/Typography';
 import { gql } from '@/gql';
 import { Screening } from '@/gql/graphql';
-import { StudentForScreening } from '@/types';
+import { InstructorScreening, MatchWithPupil, SubcourseForScreening, TutorScreening } from '@/types';
 import { SubcourseCard } from '@/widgets/course/SubcourseCard';
 import { MatchPupilCard } from '@/widgets/matching/MatchPupilCard';
 import { StudentScreeningCard } from '@/widgets/screening/StudentScreeningCard';
 import { useMutation } from '@apollo/client';
 import { useTranslation } from 'react-i18next';
 
-interface StudentHistoryProps {
-    student: StudentForScreening;
+interface StudentScreeningHistoryProps {
+    tutorScreenings: TutorScreening[];
+    instructorScreenings: InstructorScreening[];
 }
 
-export const StudentHistory = ({ student }: StudentHistoryProps) => {
+export const StudentScreeningHistory = ({ tutorScreenings, instructorScreenings }: StudentScreeningHistoryProps) => {
     const { t } = useTranslation();
     const [mutationUpdateTutorScreening] = useMutation(
         gql(`
@@ -30,9 +31,6 @@ export const StudentHistory = ({ student }: StudentHistoryProps) => {
         `)
     );
 
-    const activeMatches = student!.matches!.filter((it) => !it!.dissolved).sort((a, b) => +new Date(b!.createdAt) - +new Date(a!.createdAt));
-    const dissolvedMatches = student.matches!.filter((it) => it!.dissolved).sort((a, b) => +new Date(b!.createdAt) - +new Date(a!.createdAt));
-
     const handleOnUpdateTutorScreening = (screeningId: number, data: Pick<Screening, 'comment'>) => {
         mutationUpdateTutorScreening({ variables: { screeningId, comment: data.comment } });
     };
@@ -40,6 +38,31 @@ export const StudentHistory = ({ student }: StudentHistoryProps) => {
     const handleOnUpdateInstructorScreening = (screeningId: number, data: Pick<Screening, 'comment'>) => {
         mutationUpdateInstructorScreening({ variables: { screeningId, comment: data.comment } });
     };
+
+    return (
+        <div className="flex flex-col gap-y-2">
+            {((tutorScreenings?.length ?? 0) > 0 || (instructorScreenings?.length ?? 0) > 0) && (
+                <Typography variant="h6">{t('screening.previous_screenings')}</Typography>
+            )}
+            {tutorScreenings?.map((tutorScreening) => (
+                <StudentScreeningCard screeningType="tutor" onUpdate={handleOnUpdateTutorScreening} screening={tutorScreening} />
+            ))}
+            {instructorScreenings?.map((instructorScreening) => (
+                <StudentScreeningCard screeningType="instructor" onUpdate={handleOnUpdateInstructorScreening} screening={instructorScreening} />
+            ))}
+        </div>
+    );
+};
+
+interface StudentMatchingHistoryProps {
+    matches: MatchWithPupil[];
+}
+
+export const StudentMatchingHistory = ({ matches }: StudentMatchingHistoryProps) => {
+    const { t } = useTranslation();
+
+    const activeMatches = matches!.filter((it) => !it!.dissolved).sort((a, b) => +new Date(b!.createdAt) - +new Date(a!.createdAt));
+    const dissolvedMatches = matches!.filter((it) => it!.dissolved).sort((a, b) => +new Date(b!.createdAt) - +new Date(a!.createdAt));
 
     return (
         <div className="flex flex-col gap-y-6">
@@ -59,27 +82,20 @@ export const StudentHistory = ({ student }: StudentHistoryProps) => {
                     ))}
                 </div>
             )}
-            {student.subcoursesInstructing.length > 0 && (
-                <div className="flex flex-col gap-y-2">
-                    <Typography variant="h6">{t('screening.their_courses')}</Typography>
-                    {student.subcoursesInstructing.map((subcourse) => (
-                        <SubcourseCard subcourse={subcourse} />
-                    ))}
-                </div>
-            )}
-            {(!!student.tutorScreenings?.length || !!student.instructorScreenings?.length) && (
-                <div className="flex flex-col gap-y-2">
-                    {((student.tutorScreenings?.length ?? 0) > 0 || (student.instructorScreenings?.length ?? 0) > 0) && (
-                        <Typography variant="h6">{t('screening.previous_screenings')}</Typography>
-                    )}
-                    {student.tutorScreenings?.map((tutorScreening) => (
-                        <StudentScreeningCard screeningType="tutor" onUpdate={handleOnUpdateTutorScreening} screening={tutorScreening} />
-                    ))}
-                    {student.instructorScreenings?.map((instructorScreening) => (
-                        <StudentScreeningCard screeningType="instructor" onUpdate={handleOnUpdateInstructorScreening} screening={instructorScreening} />
-                    ))}
-                </div>
-            )}
+        </div>
+    );
+};
+
+interface StudentCourseHistoryProps {
+    subcourses: SubcourseForScreening[];
+}
+
+export const StudentCourseHistory = ({ subcourses }: StudentCourseHistoryProps) => {
+    return (
+        <div className="flex flex-col gap-y-2">
+            {subcourses.map((subcourse) => (
+                <SubcourseCard subcourse={subcourse} />
+            ))}
         </div>
     );
 };
